Extracted visitor to separate class
